+++ Network/API/Telegram/Bot/Field.hs
@@ -1,0 +1,11 @@
+module Network.API.Telegram.Bot.Field (module Exports) where
+
+import Network.API.Telegram.Bot.Field.Caption as Exports
+import Network.API.Telegram.Bot.Field.Duration as Exports
+import Network.API.Telegram.Bot.Field.Filesize as Exports
+import Network.API.Telegram.Bot.Field.Language as Exports
+import Network.API.Telegram.Bot.Field.Name as Exports
+import Network.API.Telegram.Bot.Field.MIME as Exports
+import Network.API.Telegram.Bot.Field.Height as Exports
+import Network.API.Telegram.Bot.Field.Width as Exports
+import Network.API.Telegram.Bot.Field.URI as Exports