+++ Network/API/Telegram/Bot/Object.hs
@@ -1,6 +1,5 @@
 module Network.API.Telegram.Bot.Object (module Exports) where
 
-import Network.API.Telegram.Bot.Object.Special as Exports
 import Network.API.Telegram.Bot.Object.Update as Exports
 import Network.API.Telegram.Bot.Object.Sender as Exports
 import Network.API.Telegram.Bot.Object.Member as Exports