+++ Network/API/Telegram/Bot/Object/Special.hs
@@ -1,4 +1,0 @@
-module Network.API.Telegram.Bot.Object.Special (module Exports) where
-
-import Network.API.Telegram.Bot.Object.Special.Language as Exports
-import Network.API.Telegram.Bot.Object.Special.Name as Exports