+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Document.hs
@@ -8,7 +8,7 @@ import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special (Filesize, MIME, URI)
+import Network.API.Telegram.Bot.Field (Filesize, MIME, URI)
 import Network.API.Telegram.Bot.Property (Identifiable (Identificator, ident))
 
 data Document = Document URI (Maybe Text) (Maybe MIME) (Maybe Filesize) deriving Show